add responsive sidebar
